moderation_bot/requirements.txt
python-telegram-bot[job-queue]
python-dotenv
structlog
pytest
pytest-asyncio
